Add `preserve-symlinks` option to dev server
Since we're doing all of the development with symlinks, you have to tell the server to treat them as though they are in their "real" location instead of as a symlink to another directory. Otherwise you get an error like this:
```
Error compiling: Import "@internetarchive/waveform-progress" resolved to the file "/Users/jason/Documents/Projects/iaux/packages/waveform-progress/index.js" which is outside the root directory. Install the module locally in the current project, or expand the root directory. If this is a symlink or if you used npm link,  you can use the preserveSymlinks option
```